9. Trigger network updates
