+++ app/src/main/java/com/example/android/codelabs/paging/data/GithubRepository.kt
@@ -31,15 +31,6 @@ class GithubRepository(
     private val service: GithubService,
     private val cache: GithubLocalCache
 ) {
-    // keep the last requested page. When the request is successful, increment the page number.
-    private var lastRequestedPage = 1
-
-    // LiveData of network errors.
-    private val networkErrors = MutableLiveData<String>()
-
-    // avoid triggering multiple requests in the same time
-    private var isRequestInProgress = false
-
     /**
      * Search repositories whose names match the query.
      */
@@ -49,34 +40,20 @@ class GithubRepository(
         // Get data source factory from the local cache
         val dataSourceFactory = cache.reposByName(query)
 
+        // Construct the boundary callback
+        val boundaryCallback = RepoBoundaryCallback(query, service, cache)
+        val networkErrors = boundaryCallback.networkErrors
+
         // Get the paged list
-        val data = LivePagedListBuilder(dataSourceFactory, DATABASE_PAGE_SIZE).build()
+        val data = LivePagedListBuilder(dataSourceFactory, DATABASE_PAGE_SIZE)
+                .setBoundaryCallback(boundaryCallback)
+                .build()
 
         // Get the network errors exposed by the boundary callback
         return RepoSearchResult(data, networkErrors)
     }
 
-    fun requestMore(query: String) {
-        requestAndSaveData(query)
-    }
-
-    private fun requestAndSaveData(query: String) {
-        if (isRequestInProgress) return
-
-        isRequestInProgress = true
-        searchRepos(service, query, lastRequestedPage, NETWORK_PAGE_SIZE, { repos ->
-            cache.insert(repos) {
-                lastRequestedPage++
-                isRequestInProgress = false
-            }
-        }, { error ->
-            networkErrors.postValue(error)
-            isRequestInProgress = false
-        })
-    }
-
     companion object {
-        private const val NETWORK_PAGE_SIZE = 50
         private const val DATABASE_PAGE_SIZE = 20
     }
 }
