+++ app/src/main/java/com/example/android/codelabs/paging/data/RepoBoundaryCallback.kt
@@ -1,0 +1,57 @@
+package com.example.android.codelabs.paging.data
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.paging.PagedList
+import com.example.android.codelabs.paging.api.GithubService
+import com.example.android.codelabs.paging.api.searchRepos
+import com.example.android.codelabs.paging.db.GithubLocalCache
+import com.example.android.codelabs.paging.model.Repo
+
+class RepoBoundaryCallback(
+        private val query: String,
+        private val service: GithubService,
+        private val cache: GithubLocalCache
+) : PagedList.BoundaryCallback<Repo>() {
+
+    private val _networkErrors = MutableLiveData<String>()
+
+    // keep the last requested page. When the request is successful, increment the page number.
+    private var lastRequestedPage = 1
+
+    // avoid triggering multiple requests in the same time
+    private var isRequestInProgress = false
+
+    // LiveData of network errors.
+    val networkErrors: LiveData<String>
+        get() = _networkErrors
+
+    override fun onZeroItemsLoaded() {
+        super.onZeroItemsLoaded()
+        requestAndSaveData(query)
+    }
+
+    override fun onItemAtEndLoaded(itemAtEnd: Repo) {
+        super.onItemAtEndLoaded(itemAtEnd)
+        requestAndSaveData(query)
+    }
+
+    private fun requestAndSaveData(query: String) {
+        if (isRequestInProgress) return
+
+        isRequestInProgress = true
+        searchRepos(service, query, lastRequestedPage, NETWORK_PAGE_SIZE, { repos ->
+            cache.insert(repos) {
+                lastRequestedPage++
+                isRequestInProgress = false
+            }
+        }, { error ->
+            _networkErrors.postValue(error)
+            isRequestInProgress = false
+        })
+    }
+
+    companion object {
+        private const val NETWORK_PAGE_SIZE = 50
+    }
+}
