+++ app/src/main/java/com/example/android/codelabs/paging/ui/SearchRepositoriesActivity.kt
@@ -49,7 +49,6 @@ class SearchRepositoriesActivity : AppCompatActivity() {
         // add dividers between RecyclerView's row items
         val decoration = DividerItemDecoration(this, DividerItemDecoration.VERTICAL)
         list.addItemDecoration(decoration)
-        setupScrollListener()
 
         initAdapter()
         val query = savedInstanceState?.getString(LAST_SEARCH_QUERY) ?: DEFAULT_QUERY
@@ -112,12 +111,6 @@ class SearchRepositoriesActivity : AppCompatActivity() {
         } else {
             emptyList.visibility = View.GONE
             list.visibility = View.VISIBLE
-        }
-    }
-
-    private fun setupScrollListener() {
-        RecyclerViewScrollThresholdListener(list).addOnThresholdListener(SearchRepositoriesViewModel.VISIBLE_THRESHOLD){
-            viewModel.loadMoreData()
         }
     }
