+++ app/src/main/java/com/example/android/codelabs/paging/ui/SearchRepositoriesViewModel.kt
@@ -31,10 +31,6 @@ import com.example.android.codelabs.paging.model.RepoSearchResult
  */
 class SearchRepositoriesViewModel(private val repository: GithubRepository) : ViewModel() {
 
-    companion object {
-        const val VISIBLE_THRESHOLD = 5
-    }
-
     private val queryLiveData = MutableLiveData<String>()
     private val repoResult: LiveData<RepoSearchResult> = Transformations.map(queryLiveData) {
         repository.search(it)
@@ -50,13 +46,6 @@ class SearchRepositoriesViewModel(private val repository: GithubRepository) : Vi
      */
     fun searchRepo(queryString: String) {
         queryLiveData.postValue(queryString)
-    }
-
-    fun loadMoreData() {
-        val immutableQuery = lastQueryValue()
-        if (immutableQuery != null) {
-            repository.requestMore(immutableQuery)
-        }
     }
 
     /**
